Fix driver to work with newer kernel versions
A recent Fedora kernel update (v2.6.34.6-47.fc13) broke your
samsung-backlight module due to an extra argument being added; I
managed to fix it with the attached patch.

As many Fedora HOW-TO sites list your GitHub repo as a resource, I
think it would be beneficial to share it and perhaps push it to said
repo. I've tested it on my machines (plural).